검색 리스트 db 연동

+++ testRecyclerView/app/src/main/java/org/techtown/testrecyclerview/search/FoodData.kt
@@ -1,6 +1,6 @@
 package org.techtown.testrecyclerview.search
 
-data class FoodData (var foodName: String, var calorie: Int, var amount: Int,
-                     var nutri1: Int, var nutri2: Int, var nutri3: Int){
+data class FoodData (var foodName: String, var calorie: Double, var amount: Int,
+                     var nutri1: Double, var nutri2: Double, var nutri3: Double){
 
 }
